requirements.txt
# Core dependencies
pyautogen>=0.2.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Web scraping and parsing
lxml>=4.9.0

# Audio processing (for voice features)
# Required for Whisper-based speech recognition
openai-whisper==20231117

# Optional: Real-time audio recording
# Uncomment if you want real-time voice recording
# pyaudio>=0.2.11

# Optional: Advanced TTS
# Uncomment for better text-to-speech
# pyttsx3>=2.90

# Optional: Google AI support
# Uncomment if using Google's API
# google-genai
# vertexai

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# System utilities
pathlib2>=2.3.0; python_version < "3.4"
